Convert storage system to use service tracing.
This converts the storage system service to use a service trace provider.
